Cleanup #4

By turning on more compiler warnings, some soft
parts of the client were discovered and hardened.

+++ NBClient/NBClient/NBAccount.m
@@ -279,7 +279,7 @@ static NBLogLevel LogLevel = NBLogLevelWarning;
     if (![self.authenticator.credentialIdentifier isEqualToString:[self baseURL].host]) { return; }
     self.authenticator.credentialIdentifier =
     [self.authenticator.credentialIdentifier stringByAppendingString:
-     ((self.name && self.identifier != NSNotFound) ? [NSString stringWithFormat:@"-%@-%lu", self.name, self.identifier] : @"")];
+     ((self.name && self.identifier != NSNotFound) ? [NSString stringWithFormat:@"-%@-%lu", self.name, (unsigned long)self.identifier] : @"")];
 }
 
 @end
